refactor: 상품 등록페이지 헤더, 이미지 부분 분리
- #39, #37
